% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{cfun}
\alias{cfun}
\title{MLE of a Cauchy distribution
Computes the liklihood of a Cauchy distribution}
\usage{
cfun(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
MLE of a Cauchy distribution
Computes the liklihood of a Cauchy distribution
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9)
cfun(x)
}
