% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Data from design and expirement}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
d
}
\description{
Data from design and expirement
}
\references{
\url{http://users.stat.umn.edu/~corbett/classes/5303/RDataFiles/ex8.2}

\url{http://users.stat.umn.edu/~corbett/classes/5303/RDataFiles/ex8.1}

\url{http://users.stat.umn.edu/~corbett/classes/5303/RDataFiles/ex6.1}
}
\author{
Gery Olhert \email{@umn.edu}
}
\keyword{datasets}
