% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbinom.R
\name{dFun}
\alias{dFun}
\title{MLE of a dbinom distribution
Computes the liklihood of a dbinom distribution}
\usage{
dFun(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
MLE of a dbinom distribution
Computes the liklihood of a dbinom distribution
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9)
dFun(x)
}
