% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMX.R
\name{func3}
\alias{func3}
\title{MLE of gamma distribution
Computes the liklihood of a gamma distribution}
\usage{
func3(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
MLE of gamma distribution
Computes the liklihood of a gamma distribution
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9)
func3(x)
}
