% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat_Mult.R
\name{inv}
\alias{inv}
\title{gets inverse of matrix and mulitlpies it.}
\usage{
inv(x, a)
}
\arguments{
\item{infile}{Path to the imput file}
}
\value{
A matix of the infile
}
\description{
gets inverse of matrix and mulitlpies it.
}
