% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myapply.R
\name{myapply}
\alias{myapply}
\title{This is a function version of apply
The three argumnets is the array, the margin (1 or 2), and the function}
\usage{
myapply(X, Margin, fun, ..)
}
\value{
array
}
\description{
This is a function version of apply
The three argumnets is the array, the margin (1 or 2), and the function
}
