% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotMyData}
\alias{plotMyData}
\title{Wrapper function for ggplot2 for data d
Computes the mean, variance and sd of a vector}
\usage{
plotMyData(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Wrapper function for ggplot2 for data d
Computes the mean, variance and sd of a vector
}
\examples{
d<-c(2,1,3,4,6,7,78,8,8,8,87,77,6,434,6,5)
data(d)
plotMyData(d)
}
