% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_col.R
\name{stan}
\alias{stan}
\title{Takes a numeric matrix and standardizes its columns}
\usage{
stan(x)
}
\value{
A matix
}
\description{
A matrix with only one row cannot be standardized
so all matricies must have more than one row
}
