% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simp_stat.R
\name{stts}
\alias{stts}
\title{Calculate Mean, Variane, SD (again)
Computes the mean, variance and sd of a vector, but with user checks
X must be numeric,have length greater than zero, is finite, is a number.}
\usage{
stts(x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Calculate Mean, Variane, SD (again)
Computes the mean, variance and sd of a vector, but with user checks
X must be numeric,have length greater than zero, is finite, is a number.
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9,10)
stts(x)
}
